refactor: fix lib/stores circular dependency — co-locate game stores (#377)
lib/ must be a one-way dependency layer. Three violations existed:

1. lib/stores/tetrisStore.ts imported from app/games/tetris/
2. lib/stores/buildingStore.ts + slices imported from app/games/building/
3. lib/stores/index.ts re-exported useHebrewLettersStore from app/games/

Fix — Option B (co-locate with game):

Tetris:
  - Moved lib/stores/tetrisStore.ts → app/games/tetris/store/tetrisStore.ts
  - Updated 3 callers (useTouchControls, useTetrisGame, useTetrisState)
  - Removed barrel re-export from lib/stores/index.ts

Building:
  - Moved lib/stores/buildingStore.ts → app/games/building/store/buildingStore.ts
  - Moved lib/stores/building/*.ts → app/games/building/store/
  - Updated 11 callers to @/app/games/building/store/buildingStore
  - buildingStore.ts was not in the barrel — no index.ts change needed

Hebrew Letters:
  - Removed re-export from lib/stores/index.ts
  - All callers already imported directly from the game folder

Closes #377
