+++ gamesformykids/app/games/building/components/canvas/BlockShape.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useBuildingStore } from '@/lib/stores/buildingStore';
+import { useBuildingStore } from '@/app/games/building/store/buildingStore';
 
 interface BlockShapeProps {
   blockId: string;
