+++ gamesformykids/app/games/building/components/canvas/BuildingCanvas.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useBuildingStore } from '@/lib/stores/buildingStore';
+import { useBuildingStore } from '@/app/games/building/store/buildingStore';
 import BlockRenderer from './BlockRenderer';
 import EmptyCanvasWelcome from './EmptyCanvasWelcome';
