+++ gamesformykids/app/games/building/components/controls/SettingsPanel.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Volume2, VolumeX, Save } from 'lucide-react';
-import { useBuildingStore } from '@/lib/stores/buildingStore';
+import { useBuildingStore } from '@/app/games/building/store/buildingStore';
 
 export default function SettingsPanel() {
   const soundEnabled = useBuildingStore((s) => s.soundEnabled);
